# openapi/ai-capabilities.yaml
openapi: "3.0.3"
info:
  title: AI 能力 API
  description: |
    AI 能力模块，包含：
    - AI 核心（对话、版本、行为、同步）
    - AI 健康检查
    - AI 监控
    - 插件市场
    - RAG 知识库
    - LLM Provider
    - AI 协作
    - 提示词模板
    - Agent Workflow
    - 高级功能
    - Provider 管理
    - AI 模型
    - 文档解析
  version: "1.0.0"
servers:
  - url: http://localhost:3001
    description: 开发环境
...
paths:
...
  /api/ai/industries/{industry}/templates:
    get:
      tags: [ai-core]
      summary: 行业模板
      operationId: getIndustryTemplates
      parameters:
        - name: industry
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        description:
                          type: string
...
  /api/ai/sync/schema/{schemaId}/flows:
    get:
      tags: [ai-core]
      summary: Schema → Flow 反向查询
      operationId: getSchemaRelatedFlows
      parameters:
        - name: schemaId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    workflowKeyAuth:
      type: apiKey
      in: header
      name: X-Workflow-Key